Fix AIL_digital_CPU_percent / AIL_background_CPU_percent return type (S32, was F32)
Both 6.1 and 9.3 SDK headers declare these as S32 (returned in EAX), but we
declared callconv(.winapi) f32 (returned in ST(0)). A game reading the integer
percentage from EAX would get garbage. Return i32, converting the internal
float estimate via @intFromFloat. @N decoration (and thus export parity) is
unchanged since return type doesn't affect stdcall arg-byte count. 221 tests
pass; v6 stays byte-exact.

+++ src/api/v7.zig
@@ -318,7 +318,8 @@ pub fn AIL_find_filter(name: [*:0]const u8, ret: ?*?*openmiles.Provider) callcon
     return if (openmiles.startup_provider != null) 1 else 0;
 }
 
-pub fn AIL_background_CPU_percent() callconv(.winapi) f32 {
+// SDK returns S32 (integer percent in EAX), not F32.
+pub fn AIL_background_CPU_percent() callconv(.winapi) i32 {
     return 0;
 }
